Rename the IsoSpec.cpp/.h files to IsoSpecWrapper.cpp/.h to match the name of classes inside
