+++ src/tests/class_tests/openms/source/IsoSpec_test.cpp
@@ -36,7 +36,7 @@
 #include <OpenMS/test_config.h>
 
 ///////////////////////////
-#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/IsoSpec.h>
+#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/IsoSpecWrapper.h>
 ///////////////////////////
 
 #include <OpenMS/CHEMISTRY/Element.h>
